backend/requirements.txt
Flask==2.0.1
Flask-CORS==6.0.1
requests==2.26.0
google-generativeai==0.7.1
python-dotenv==0.19.0
Werkzeug==2.0.3
SQLAlchemy 